+++ shoop_tests/functional/test_tax_system.py
@@ -26,7 +26,7 @@ TAX_MODULE_SPEC = [__name__ + ":IrvineCaliforniaTaxation"]
 class IrvineCaliforniaTaxation(TaxModule):
     identifier = "irvine"
 
-    def get_taxed_price_for(self, context, item, price):
+    def get_taxed_price(self, context, price, tax_class):
         taxes = []
         if context.postal_code == "92602":
             taxes = [
@@ -37,6 +37,10 @@ class IrvineCaliforniaTaxation(TaxModule):
                 get_tax("CA-OC-IR-DS", "District tax", rate="0.005"),
             ]
         return stacked_value_added_taxes(price, taxes)
+
+    def _add_proportional_taxes(self, context, tax_class_proportions, lines):
+        for line in lines:
+            line.taxes = self.get_taxed_price(context, line.price, None).taxes
 
 
 @pytest.mark.django_db